Split WorkflowManager in three classes
