+++ src/App/src/Forms/MainForm.Init.cs
@@ -432,8 +432,8 @@ namespace mrHelper.App.Forms
       {
          _workflowManager = new WorkflowManager(Program.Settings);
          _expressionResolver = new ExpressionResolver(_workflowManager);
-         _mergeRequestCache = new MergeRequestCache(_workflowManager, _workflowManager, this, Program.Settings,
-            Program.Settings.AutoUpdatePeriodMs);
+         _mergeRequestCache = new MergeRequestCache(_workflowManager, _workflowManager, _workflowManager,
+            this, Program.Settings, Program.Settings.AutoUpdatePeriodMs);
          _discussionManager = new DiscussionManager(Program.Settings, _workflowManager, _mergeRequestCache,
             this, _keywords, Program.Settings.AutoUpdatePeriodMs, _mergeRequestFilter);
          _eventFilter = new EventFilter(Program.Settings, _workflowManager, _mergeRequestCache, _mergeRequestFilter);
