+++ src/App/src/Helpers/Git/GitStatisticManager.cs
@@ -30,6 +30,7 @@ namespace mrHelper.App.Helpers
          IProjectUpdateContextProviderFactory contextProviderFactory)
       {
          _workflowEventNotifier = workflowEventNotifier;
+         _workflowEventNotifier.Connecting += onConnecting;
          _workflowEventNotifier.Connected += onConnected;
 
          _factoryAccessor = factoryAccessor;
@@ -40,6 +41,7 @@ namespace mrHelper.App.Helpers
 
       public void Dispose()
       {
+         _workflowEventNotifier.Connecting -= onConnecting;
          _workflowEventNotifier.Connected -= onConnected;
 
          unsubscribeFromAll();
@@ -102,12 +104,10 @@ namespace mrHelper.App.Helpers
       async private Task updateGitStatistic(ILocalGitRepository repo)
       {
          Debug.Assert(isConnected(repo));
-
-         ILocalGitRepositoryData data = repo.Data;
-         if (data == null)
+         if (repo.Data == null || repo.ExpectingClone)
          {
             Trace.TraceWarning(String.Format(
-               "[GitStatisticManager] Update failed. LocalGitRepositoryData is not ready (Host={0}, Project={1})",
+               "[GitStatisticManager] Update failed. Repository is not ready (Host={0}, Project={1})",
                repo.ProjectKey.HostName, repo.ProjectKey.ProjectName));
             return;
          }
@@ -119,11 +119,12 @@ namespace mrHelper.App.Helpers
 
          try
          {
-            if (!repo.ExpectingClone)
+            await repo.Updater.SilentUpdate(getContextProvider(repo));
+            if (isConnected(repo))
             {
-               await repo.Updater.SilentUpdate(getContextProvider(repo));
+               // LocalGitRepository might be removed from collection while we were updating
+               await doUpdateGitStatistic(repo);
             }
-            await doUpdateGitStatistic(repo);
          }
          finally
          {
@@ -302,8 +303,14 @@ namespace mrHelper.App.Helpers
          unsubscribeFromOne(repo);
       }
 
-      private void onConnected(string hostname, IEnumerable<Project> projects)
+      private void onConnecting(string hostname)
       {
+         unsubscribeFromAll();
+      }
+
+      private void onConnected(string hostname, User user, IEnumerable<Project> projects)
+      {
+         Debug.Assert(!_gitStatistic.Any());
          foreach (Project project in projects)
          {
             ProjectKey key = new ProjectKey
