+++ src/GitLabClient/src/Workflow/WorkflowManager.cs
@@ -1,90 +1,40 @@
 using System;
-using System.Linq;
-using System.Diagnostics;
 using System.Threading.Tasks;
 using System.Collections.Generic;
-using GitLabSharp;
 using GitLabSharp.Entities;
 using GitLabSharp.Accessors;
-using Version = GitLabSharp.Entities.Version;
 using mrHelper.Client.Types;
 using mrHelper.Client.Common;
 using mrHelper.Common.Tools;
 using mrHelper.Common.Interfaces;
-using mrHelper.Common.Exceptions;
 using mrHelper.Common.Constants;
 
 namespace mrHelper.Client.Workflow
 {
-   public class WorkflowException : ExceptionEx
+   public class WorkflowManager : BaseWorkflowManager, IWorkflowEventNotifier, IMergeRequestListLoader
    {
-      internal WorkflowException(string message, Exception innerException)
-         : base(message, innerException) {}
+      public event Action<string> Connecting;
+      public event Action<string, User, IEnumerable<Project>> Connected;
 
-      public string UserMessage
-      {
-         get
-         {
-            if (InnerException is OperatorException ox)
-            {
-               if (ox.InnerException is GitLabRequestException rx)
-               {
-                  if (rx.InnerException is System.Net.WebException wx)
-                  {
-                     System.Net.HttpWebResponse response = wx.Response as System.Net.HttpWebResponse;
-                     if (response != null && response.StatusCode == System.Net.HttpStatusCode.Unauthorized)
-                     {
-                        return wx.Message + " Check your access token!";
-                     }
-                     return wx.Message;
-                  }
-               }
-            }
-            return OriginalMessage;
-         }
-      }
-   }
-
-   /// <summary>
-   /// Provides access to main GitLab workflow actions: load user, load merge requests etc
-   /// Supports chains of actions (loading a merge request also loads its versions or commits)
-   /// Each action toggles Pre-{Action}-Event and either Post-{Action}-Event or Failed-{Action}-Event
-   /// </summary>
-   public class WorkflowManager : IWorkflowEventNotifier, IWorkflowLoader
-   {
-      public event Action<string, User> Connecting;
-      public event Action<string, IEnumerable<Project>> Connected;
+      public event Action<Project> PreLoadProjectMergeRequests;
+      public event Action<string, Project, IEnumerable<MergeRequest>> PostLoadProjectMergeRequests;
+      public event Action FailedLoadProjectMergeRequests;
 
       public WorkflowManager(IHostProperties settings)
+         : base(settings)
       {
-         _settings = settings;
-      }
-
-      async public Task<bool> LoadCurrentUserAsync(string hostname)
-      {
-         _operator = new WorkflowDataOperator(hostname, _settings.GetAccessToken(hostname));
-
-         return await loadCurrentUserAsync(hostname);
-      }
-
-      async public Task<bool> LoadAllMergeRequestsAsync(string hostname, object search, int? maxResults)
-      {
-         _operator = new WorkflowDataOperator(hostname, _settings.GetAccessToken(hostname));
-
-         Dictionary<Project, IEnumerable<MergeRequest>> mergeRequests =
-            await loadMergeRequestsAsync(hostname, search, maxResults);
-         if (mergeRequests == null)
-         {
-            return false; // cancelled
-         }
-
-         return true;
       }
 
       async public Task<bool> LoadAllMergeRequestsAsync(string hostname, IEnumerable<Project> projects,
          Action<string, string> onForbiddenProject, Action<string, string> onNotFoundProject)
       {
          _operator = new WorkflowDataOperator(hostname, _settings.GetAccessToken(hostname));
+
+         User? currentUser = await loadCurrentUserAsync(hostname);
+         if (!currentUser.HasValue)
+         {
+            return false;
+         }
 
          Exception exception = null;
          bool cancelled = false;
@@ -146,11 +96,12 @@ namespace mrHelper.Client.Workflow
             }
          }
 
+         Connecting?.Invoke(hostname);
          await TaskUtils.RunConcurrentFunctionsAsync(projects, x => loadProject(x),
             Constants.ProjectsInBatch, Constants.ProjectsInterBatchDelay, () => cancelled);
          if (!cancelled)
          {
-            Connected?.Invoke(hostname, projects);
+            Connected?.Invoke(hostname, currentUser.Value, projects);
             return true;
          }
 
@@ -161,62 +112,19 @@ namespace mrHelper.Client.Workflow
          return false;
       }
 
-      async public Task<bool> LoadMergeRequestAsync(string hostname, string projectname, int mergeRequestIId,
-         EComparableEntityType comparableEntityType)
+      async private Task<User?> loadCurrentUserAsync(string hostName)
       {
-         _operator = new WorkflowDataOperator(hostname, _settings.GetAccessToken(hostname));
-
-         return await loadMergeRequestAsync(hostname, projectname, mergeRequestIId, comparableEntityType);
-      }
-
-      async public Task CancelAsync()
-      {
-         if (_operator != null)
-         {
-            await _operator.CancelAsync();
-         }
-      }
-
-      public event Action<string> PreLoadCurrentUser;
-      public event Action<string, User> PostLoadCurrentUser;
-      public event Action FailedLoadCurrentUser;
-
-      public event Action<Project> PreLoadProjectMergeRequests;
-      public event Action<string, Project, IEnumerable<MergeRequest>> PostLoadProjectMergeRequests;
-      public event Action FailedLoadProjectMergeRequests;
-
-      public event Action<int> PreLoadSingleMergeRequest;
-      public event Action<string, string, MergeRequest> PostLoadSingleMergeRequest;
-      public event Action FailedLoadSingleMergeRequest;
-
-      public event Action PreLoadComparableEntities;
-      public event Action<string, string, MergeRequest, System.Collections.IEnumerable> PostLoadComparableEntities;
-      public event Action FailedLoadComparableEntities;
-
-      public event Action PreLoadVersions;
-      public event Action<string, string, MergeRequest, IEnumerable<Version>> PostLoadVersions;
-      public event Action FailedLoadVersions;
-
-      async private Task<bool> loadCurrentUserAsync(string hostName)
-      {
-         PreLoadCurrentUser?.Invoke(hostName);
-
-         User currentUser;
          try
          {
-            currentUser = await _operator.GetCurrentUserAsync();
+            return await _operator.GetCurrentUserAsync();
          }
          catch (OperatorException ex)
          {
             string cancelMessage = String.Format("Cancelled loading current user from host \"{0}\"", hostName);
             string errorMessage = String.Format("Cannot load user from host \"{0}\"", hostName);
-            handleOperatorException(ex, cancelMessage, errorMessage, new Action[] { FailedLoadCurrentUser });
-            return false;
+            handleOperatorException(ex, cancelMessage, errorMessage, null);
          }
-
-         PostLoadCurrentUser?.Invoke(hostName, currentUser);
-         Connecting?.Invoke(hostName, currentUser);
-         return true;
+         return null;
       }
 
       async private Task<IEnumerable<MergeRequest>> loadProjectMergeRequestsAsync(string hostname, Project project)
@@ -243,198 +151,6 @@ namespace mrHelper.Client.Workflow
          return mergeRequests;
       }
 
-      async private Task<Dictionary<Project, IEnumerable<MergeRequest>>> loadMergeRequestsAsync(
-         string hostname, object search, int? maxResults)
-      {
-         IEnumerable<MergeRequest> mergeRequests;
-         try
-         {
-            mergeRequests = await _operator.SearchMergeRequestsAsync(search, maxResults, false);
-         }
-         catch (OperatorException ex)
-         {
-            string cancelMessage = String.Format("Cancelled loading merge requests with search string \"{0}\"", search);
-            string errorMessage = String.Format("Cannot load merge requests with search string \"{0}\"", search);
-            handleOperatorException(ex, cancelMessage, errorMessage, new Action[] { FailedLoadProjectMergeRequests });
-            return null;
-         }
-
-         Dictionary<Project, IEnumerable<MergeRequest>> result = new Dictionary<Project, IEnumerable<MergeRequest>>();
-
-         bool cancelled = false;
-         async Task resolve(KeyValuePair<int, List<MergeRequest>> keyValuePair)
-         {
-            if (cancelled)
-            {
-               return;
-            }
-
-            Project? project = await resolveProject(hostname, keyValuePair.Key);
-            if (project == null)
-            {
-               cancelled = true;
-               return;
-            }
-            result.Add(project.Value, keyValuePair.Value);
-         }
-
-         await TaskUtils.RunConcurrentFunctionsAsync(groupMergeRequestsByProject(mergeRequests), x => resolve(x),
-            Constants.ProjectsInBatch, Constants.ProjectsInterBatchDelay, () => cancelled);
-         if (cancelled)
-         {
-            return null;
-         }
-
-         foreach (KeyValuePair<Project, IEnumerable<MergeRequest>> keyValuePair in result)
-         {
-            PreLoadProjectMergeRequests?.Invoke(keyValuePair.Key);
-            PostLoadProjectMergeRequests?.Invoke(hostname, keyValuePair.Key, keyValuePair.Value);
-         }
-
-         return result;
-      }
-
-      private Dictionary<int, List<MergeRequest>> groupMergeRequestsByProject(IEnumerable<MergeRequest> mergeRequests)
-      {
-         Dictionary<int, List<MergeRequest>> grouped = new Dictionary<int, List<MergeRequest>>();
-         foreach (MergeRequest mergeRequest in mergeRequests)
-         {
-            int projectId = mergeRequest.Project_Id;
-            if (!grouped.ContainsKey(projectId))
-            {
-               grouped[projectId] = new List<MergeRequest>();
-            }
-            grouped[projectId].Add(mergeRequest);
-         }
-         return grouped;
-      }
-
-      async private Task<Project?> resolveProject(string hostname, int projectId)
-      {
-         try
-         {
-            return await _operator.GetProjectAsync(projectId.ToString());
-         }
-         catch (OperatorException ex)
-         {
-            string cancelMessage = String.Format("Cancelled resolving project with Id \"{0}\"", projectId);
-            string errorMessage = String.Format("Cannot load project with Id \"{0}\"", projectId);
-            handleOperatorException(ex, cancelMessage, errorMessage, new Action[] { FailedLoadProjectMergeRequests });
-            return null;
-         }
-      }
-
-      async private Task<bool> loadMergeRequestAsync(string hostname, string projectName, int mergeRequestIId,
-         EComparableEntityType comparableEntityType)
-      {
-         PreLoadSingleMergeRequest?.Invoke(mergeRequestIId);
-
-         MergeRequest mergeRequest = new MergeRequest();
-         try
-         {
-            SearchByIId searchByIId = new SearchByIId { ProjectName = projectName, IId = mergeRequestIId };
-            IEnumerable<MergeRequest> mergeRequests = await _operator.SearchMergeRequestsAsync(searchByIId, null, true);
-            mergeRequest = mergeRequests.FirstOrDefault();
-         }
-         catch (OperatorException ex)
-         {
-            string cancelMessage = String.Format("Cancelled loading MR with IId {0}", mergeRequestIId);
-            string errorMessage = String.Format("Cannot load merge request with IId {0}", mergeRequestIId);
-            handleOperatorException(ex, cancelMessage, errorMessage, new Action[] { FailedLoadSingleMergeRequest });
-            return false;
-         }
-
-         PostLoadSingleMergeRequest?.Invoke(hostname, projectName, mergeRequest);
-
-         switch (comparableEntityType)
-         {
-            case EComparableEntityType.Commit:
-               return await loadVersionsAsync(hostname, projectName, mergeRequest, false)
-                   && await loadCommitsAsync(hostname, projectName, mergeRequest);
-
-            case EComparableEntityType.Version:
-               return await loadVersionsAsync(hostname, projectName, mergeRequest, true);
-         }
-
-         Debug.Assert(false);
-         return true;
-      }
-
-      async private Task<bool> loadCommitsAsync(string hostname, string projectName, MergeRequest mergeRequest)
-      {
-         PreLoadComparableEntities?.Invoke();
-         IEnumerable<Commit> commits;
-         try
-         {
-            commits = await _operator.GetCommitsAsync(projectName, mergeRequest.IId);
-         }
-         catch (OperatorException ex)
-         {
-            string cancelMessage = String.Format("Cancelled loading commits for merge request with IId {0}",
-               mergeRequest.IId);
-            string errorMessage = String.Format("Cannot load commits for merge request with IId {0}",
-               mergeRequest.IId);
-            handleOperatorException(ex, cancelMessage, errorMessage, new Action[] { FailedLoadComparableEntities });
-            return false;
-         }
-         PostLoadComparableEntities?.Invoke(hostname, projectName, mergeRequest, commits);
-         return true;
-      }
-
-      async private Task<bool> loadVersionsAsync(string hostname, string projectName, MergeRequest mergeRequest,
-         bool invokeCompareableEntitiesCallback)
-      {
-         List<Action> failureActions = new List<Action> { FailedLoadVersions };
-         if (invokeCompareableEntitiesCallback)
-         {
-            failureActions.Add(FailedLoadComparableEntities);
-         }
-
-         if (invokeCompareableEntitiesCallback)
-         {
-            PreLoadComparableEntities?.Invoke();
-         }
-         PreLoadVersions?.Invoke();
-
-         IEnumerable<Version> versions;
-         try
-         {
-            versions = await _operator.GetVersionsAsync(projectName, mergeRequest.IId);
-         }
-         catch (OperatorException ex)
-         {
-            string cancelMessage = String.Format("Cancelled loading versions for merge request with IId {0}",
-               mergeRequest.IId);
-            string errorMessage = String.Format("Cannot load versions for merge request with IId {0}",
-               mergeRequest.IId);
-            handleOperatorException(ex, cancelMessage, errorMessage, failureActions);
-            return false;
-         }
-
-         if (invokeCompareableEntitiesCallback)
-         {
-            PostLoadComparableEntities?.Invoke(hostname, projectName, mergeRequest, versions);
-         }
-         PostLoadVersions?.Invoke(hostname, projectName, mergeRequest, versions);
-
-         return true;
-      }
-
-      private void handleOperatorException(OperatorException ex, string cancelMessage, string errorMessage,
-         IEnumerable<Action> failureActions)
-      {
-         bool cancelled = ex.InnerException is GitLabClientCancelled;
-         if (cancelled)
-         {
-            Trace.TraceInformation(String.Format("[WorkflowManager] {0}", cancelMessage));
-            return;
-         }
-
-         failureActions?.ToList().ForEach(x => x?.Invoke());
-
-         throw new WorkflowException(errorMessage, ex);
-      }
-
       private static bool isForbiddenProjectException(WorkflowException ex)
       {
          System.Net.HttpWebResponse response = getWebResponse(ex);
@@ -459,9 +175,6 @@ namespace mrHelper.Client.Workflow
          }
          return null;
       }
-
-      private readonly IHostProperties _settings;
-      private WorkflowDataOperator _operator;
    }
 }
